Cyndaquil Line Electric
I've changed the Cyndaquil Line from Fire/Ghost to Fire/Electric.
Updating Movesets, and trainer parties that use the team.
As well as shiny pallets to reflect a full yellow color.

Also modified Scizor's Shiny Pal to be a brighter green, and Scyther's Shiny Pal to be a deep Red (reverse of their default colours).
Modified Murkrow's shiny pal to be white like Ninetales
and modified MewTwo's Shiny Pal to be Blue like Mew.

Also updated a minor text inconsistancy with Chuck's Wife (making it clearer that she's a trainer).