Ancient Bowl
Amended to fossil revival requirements for the Amaranth Rib to now require an ancient bowl.
Also modified the Aerodactyl Item Room to contain an Old Amber, and modified the Kabuto Room to no-longer contain an Ancient Relic, but to instead contain 7 sherds of Tanoby Wear (the pot smashed in this chamber).