Splash a Pokedex Entry
Whenever you encounter a Pokemon sprite (or solve a Ruins of ALPH Puzzle) it splashes up a seen Pokedex Entry like in Gen I.
This doesn't happen for Pokemon sprites you can battle, as these Pokemon will be entered into the Pokedex the normal way.

This is arguably most useful for Entei and Raikou.