Moved Berry Juicer
Swapped the Poke-Seer in Cianwood with the Berry Juicer in the Hidden Power house.

The Poke-Seer's house is right by a bunch of bounders. It makes sense she'd have a shuckle. Meanwhile the Poke-Seer is pretty meaningless, and I think it's fine to move it to that otherwise pretty obscure house.