Coin Trader
Removed the Happiness Reader NPC (as all Pokemon display happiness on the 4th stat page)
Replaced this with a NPC that will trade Gold Leafs for 100 Coins and Silver Leafs for 50 Coins.

Also referenced this in the dialogue of the Goldenrod game corner.
And also changed the text of the TM prize Lady to reflect the TM name rather than numbers.