Weather Ball
Replaced the Move Frustration with Weather Ball!
It's an 80 base power move that has a 40% chance to Burn, Confuse, Flinch, Freeze, or Poison the target dependent on the weather.
As I replaced the move Frustration, I've made it a TM that is learnable by basically every Pokemon. Which does give some versatility to any party.