Surf and Waterfall Power Swap
I've downgraded Surf to a 75 base power move considering how early in the game it's obtained. And bumped waterfall up to a 95 base power move. Though it's still a flinch it from previous edits. This edit is only to swap out hard-coded trainer moves to make it so powerful trainers use the more powerful water-type move.