Weather Ball AI
Updated the move AI for each of the 5 weather effects to set it so that way Weather Ball is a highly recommended move. Also gave weather ball to trainers where there is guaranteed weather on the map that would align with a Pokemon's STAB
Also altered the cost and location certain TMs, and corrected an oversight with the overworld Rattata sprites.