Expand Goldenrod Archives
Added GOLDENROD_ARCHIVES1F & GOLDENROD_ARCHIVESB1F to the game.
These maps are based on the Burned Tower maps from Pokegold.
The trainers and events from the original Goldenrod Archives have been moved so that they are divided among the new maps.
The original map based on the first floor of the slowpoke well, has had the Strength boulder restored and the lower floors are accessible through the bottom of the archive.

Wild pokemon cause the game to freeze on all maps except for the two new maps. It is unclear as to why.